refactor(tab-manager): consolidate composite-id into workspace module
The composite ID types, constructors, and parsers were duplicated between
composite-id.ts and workspace.ts. Since workspace.ts is the single source
of truth for the tab manager schema, composite IDs belong there. Delete
the standalone file and update the one remaining import in FlatTabList.
